# .scratch/source/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Casino Admin Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deposits:
    get:
      operationId: listDeposits
      tags: [deposits]
      summary: List deposit requests
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected, completed]
      responses:
        "200":
          description: List of deposit requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepositRequest"
  /deposits/{id}/approve:
    patch:
      operationId: approveDeposit
      tags: [deposits]
      summary: Approve a deposit request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Updated deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositRequest"
  /deposits/{id}/reject:
    patch:
      operationId: rejectDeposit
      tags: [deposits]
      summary: Reject a deposit request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Updated deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositRequest"
  /deposits/{id}/complete:
    patch:
      operationId: completeDeposit
      tags: [deposits]
      summary: Mark deposit as completed
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Updated deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositRequest"

  # ─── REDEEM REQUESTS ──────────────────────────────────────────────────────────
...
  /game-accounts:
    get:
      operationId: listGameAccountRequests
      tags: [game-accounts]
      summary: List game account requests
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, fulfilled, rejected]
      responses:
        "200":
          description: List of game account requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameAccountRequest"
  /game-accounts/{id}/fulfill:
    patch:
      operationId: fulfillGameAccountRequest
      tags: [game-accounts]
      summary: Fulfill a game account request and send credentials
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameAccountCredentials"
      responses:
        "200":
          description: Fulfilled request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameAccountRequest"
  /game-accounts/{id}/reject:
    patch:
      operationId: rejectGameAccountRequest
      tags: [game-accounts]
      summary: Reject a game account request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameAccountRequest"

  # ─── FREE PLAY REQUESTS ───────────────────────────────────────────────────────
  /free-play:
    get:
      operationId: listFreePlayRequests
      tags: [free-play]
      summary: List free play requests
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: List of free play requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FreePlayRequest"
  /free-play/{id}/approve:
    patch:
      operationId: approveFreePlay
      tags: [free-play]
      summary: Approve free play with amount
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreePlayApproval"
      responses:
        "200":
          description: Approved free play request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreePlayRequest"
  /free-play/{id}/reject:
    patch:
      operationId: rejectFreePlay
      tags: [free-play]
      summary: Reject free play request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected free play
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreePlayRequest"

  # ─── GAMES MANAGEMENT ─────────────────────────────────────────────────────────
...
  /faqs:
    get:
      operationId: listFaqs
      tags: [faqs]
      summary: List all FAQs
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: List of FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faq"
    post:
      operationId: createFaq
      tags: [faqs]
      summary: Create a FAQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqInput"
      responses:
        "201":
          description: Created FAQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
  /faqs/{id}:
    patch:
      operationId: updateFaq
      tags: [faqs]
      summary: Update a FAQ
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqUpdate"
      responses:
        "200":
          description: Updated FAQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
    delete:
      operationId: deleteFaq
      tags: [faqs]
      summary: Delete a FAQ
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── TELEGRAM BUTTONS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RejectInput:
      type: object
      properties:
        reason:
          type: string
      required: []

    # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
    DepositRequest:
      type: object
      properties:
        id:
          type: number
        customerName:
          type: string
        customerTelegramId:
          type: string
        conversationId:
          type: number
        amount:
          type: ["number", "null"]
        paymentMethod:
          type: ["string", "null"]
        screenshotUrl:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, completed]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, customerName, customerTelegramId, conversationId, status, createdAt, updatedAt]

    # ─── REDEEMS ──────────────────────────────────────────────────────────────────
...
    GameAccountRequest:
      type: object
      properties:
        id:
          type: number
        customerName:
          type: string
        customerTelegramId:
          type: string
        conversationId:
          type: number
        gameId:
          type: number
        gameName:
          type: string
        gameLink:
          type: ["string", "null"]
        credentialsUsername:
          type: ["string", "null"]
        credentialsPassword:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, fulfilled, rejected]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, customerName, customerTelegramId, conversationId, gameId, gameName, status, createdAt, updatedAt]
    GameAccountCredentials:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        gameLink:
          type: string
      required: [username, password, gameLink]

    # ─── FREE PLAY ────────────────────────────────────────────────────────────────
    FreePlayRequest:
      type: object
      properties:
        id:
          type: number
        customerName:
          type: string
        customerTelegramId:
          type: string
        conversationId:
          type: number
        requestedAmount:
          type: ["number", "null"]
        approvedAmount:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, customerName, customerTelegramId, conversationId, status, createdAt, updatedAt]
    FreePlayApproval:
      type: object
      properties:
        amount:
          type: number
      required: [amount]

    # ─── GAMES ────────────────────────────────────────────────────────────────────
...
    Faq:
      type: object
      properties:
        id:
          type: number
        category:
          type: string
        question:
          type: string
        answer:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, category, question, answer, createdAt, updatedAt]
    FaqInput:
      type: object
      properties:
        category:
          type: string
        question:
          type: string
        answer:
          type: string
      required: [category, question, answer]
    FaqUpdate:
      type: object
      properties:
        category:
          type: string
        question:
          type: string
        answer:
          type: string

    # ─── TELEGRAM BUTTONS ─────────────────────────────────────────────────────────
...