# .scratch/source/lib/api-zod/src/generated/types/faqInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Casino Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface FaqInput {
  category: string;
  question: string;
  answer: string;
}