# .scratch/source/lib/api-zod/src/generated/types/gameAccountRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Casino Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { GameAccountRequestStatus } from './gameAccountRequestStatus';

export interface GameAccountRequest {
  id: number;
  customerName: string;
  customerTelegramId: string;
  conversationId: number;
  gameId: number;
  gameName: string;
  /** @nullable */
  gameLink?: string | null;
  /** @nullable */
  credentialsUsername?: string | null;
  /** @nullable */
  credentialsPassword?: string | null;
  status: GameAccountRequestStatus;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: Date;
  updatedAt: Date;
}